# backend/swagger.yaml
openapi: 3.0.0
info:
  title: DWF Helpdesk API
  description: |
    # DWF Helpdesk System API
    
    ระบบ Helpdesk สำหรับกรมกิจการสตรีและสถาบันครอบครัว (DWF)
    
    ## 🎯 Overview
    This API provides endpoints for managing support tickets, user authentication, and administrative functions for the DWF Helpdesk System.
    
    ## 🔐 Authentication
    Most endpoints require JWT authentication. Include the token in the Authorization header:
    `Authorization: Bearer <your_access_token>`
    
    ## 📊 Response Format
    All API responses follow this standard format:
    ```json
    {
      "success": boolean,
      "message": "string (optional)",
      "data": "object/array (optional)",
      "error": "string (optional, when success is false)"
    }
    ```
  version: 1.0.0
  contact:
    name: DWF IT Support
    email: support@dwf.go.th
servers:
  - url: http://localhost:3001/api
    description: Development Server (Local)
  - url: http://localhost:3002/api
    description: Development Server (Docker)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Authorization header using the Bearer scheme
  schemas:
...
    Ticket:
      type: object
      properties:
        id:
          type: integer
          example: 1
        ticketId:
          type: string
          example: TK1699234567890
        problemType:
          type: string
          example: ฮาร์ดแวร์
        problemDescription:
          type: string
          example: คอมพิวเตอร์เปิดไม่ติด
        fullName:
          type: string
          example: นายสมชาย ใจดี
        phoneNumber:
          type: string
          example: 0812345678
        department:
          type: string
          example: กลุ่มงานเทคโนโลยี
        division:
          type: string
          example: ฝ่ายพัฒนาระบบ
        assetNumber:
          type: string
          example: DWF-PC-001
        status:
          type: string
          example: รอดำเนินการ
        priority:
          type: string
          example: ปกติ
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        assignedTo:
          type: object
          properties:
            id:
              type: integer
            fullName:
              type: string
    CreateTicketRequest:
      type: object
      required:
        - problemType
        - problemDescription
        - fullName
        - phoneNumber
        - department
      properties:
        problemType:
          type: string
          description: ประเภทของปัญหา
          example: ฮาร์ดแวร์
        otherProblemType:
          type: string
          description: ประเภทปัญหาอื่นๆ (ถ้าเลือก "อื่นๆ")
        problemDescription:
          type: string
          description: รายละเอียดปัญหา
          example: คอมพิวเตอร์เปิดไม่ติด
        fullName:
          type: string
          description: ชื่อ-นามสกุล
          example: นายสมชาย ใจดี
        phoneNumber:
          type: string
          description: เบอร์โทรศัพท์
          example: 0812345678
        department:
          type: string
          description: กอง/กลุ่ม
          example: กลุ่มงานเทคโนโลยี
        division:
          type: string
          description: กลุ่ม/ฝ่าย
          example: ฝ่ายพัฒนาระบบ
        assetNumber:
          type: string
          description: เลขครุภัณฑ์
          example: DWF-PC-001
    TicketResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: สร้าง ticket เรียบร้อย
        ticket:
          $ref: '#/components/schemas/Ticket'
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Error message
        details:
          type: array
          items:
            type: object
...
paths:
...
  /tickets:
    post:
      tags:
        - Tickets
      summary: Create new ticket
      description: สร้าง ticket ใหม่ (ไม่ต้อง authentication)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTicketRequest'
          multipart/form-data:
            schema:
              type: object
              allOf:
                - $ref: '#/components/schemas/CreateTicketRequest'
                - type: object
                  properties:
                    attachments:
                      type: array
                      items:
                        type: string
                        format: binary
                      description: ไฟล์แนบ (สูงสุด 5 ไฟล์, 10MB ต่อไฟล์)
      responses:
        '201':
          description: สร้าง ticket สำเร็จ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TicketResponse'
        '400':
          description: ข้อมูลไม่ถูกต้อง
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: เกิน rate limit (3 tickets ต่อชั่วโมง)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /dashboard/tickets:
    get:
      tags:
        - Dashboard
      summary: Get recent tickets
      description: ดึงรายการ tickets ล่าสุดสำหรับ dashboard (Support+ only)
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          description: จำนวน tickets ที่ต้องการ
      responses:
        '200':
          description: รายการ tickets ล่าสุด
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  tickets:
                    type: array
                    items:
                      $ref: '#/components/schemas/Ticket'
        '401':
          description: ไม่ได้รับอนุญาต
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # System Endpoints
...